python/requirements.txt
cx-Oracle==8.3.0
click>=8.0
pytest
cheroot==8.5.2
flask_restx==1.3.0
configobj~=5.0.6
Flask_SQLAlchemy==3.1.0
cryptography~=41.0.5
pandas
hypothesis~=5.48.0
parsimonious~=0.8.1
SQLAlchemy==2.0.16
argcomplete~=1.12.2
cytoolz~=0.10.1
six~=1.15.0
attrdict~=2.0.1
Flask~=2.2.5
pymitter~=0.3.0
requests~=2.31.0
requests_toolbelt==0.9.1
lru_dict==1.1.6
promise~=2.3
#protobuf==3.19.0
# protobuf>=4.21.6,<5.0dev
protobuf>=5.27.1
pycryptodome==3.9.9
pyjwt
pyyaml
mysqlclient==2.1.0
waitress==2.1.2
sqlparse~=0.4.1
ecdsa==0.19.0
toolz~=0.11.1
tenacity==7.0.0
coincurve~=13.0.0
google~=3.0.0
paste~=3.5.0
func_timeout==4.3.0
cheroot==8.5.2
prefect==1.4.0
gmssl~=3.2.1
readerwriterlock~=1.0.4
jsoncomment~=0.2.3
matplotlib~=3.2.2
seaborn~=0.10.1
sqlvalidator==0.0.17
requests-toolbelt==0.9.1
hdfs
scikit-learn~=0.24.2
gmpy2
openpyxl
networkx
pydot
snowland-smx
numpy==1.23.1
graphviz
grpcio==1.62.1
grpcio-tools==1.62.1
xlrd~=1.0.0
MarkupSafe>=2.1.1
urllib3==1.26.18
phe
chardet
